Add sliding window, more cfgs

+++ mistralrs-core/src/layers.rs
@@ -1912,7 +1912,7 @@ impl GPTOSSRotaryEmbedding {
 
         let inv_freq: Vec<_> = (0..dim)
             .step_by(2)
-            .map(|i| 1f32 / cfg.rope_theta.powf(i as f64 / dim as f64) as f32)
+            .map(|i| 1f32 / cfg.rope_theta.powf(i as f32 / dim as f32) as f32)
             .collect();
         let inv_freq_len = inv_freq.len();
         let inv_freq = Tensor::from_vec(inv_freq, (1, inv_freq_len), dev)?;
